snowsql-call/requirements.txt
snowflake-connector-python==3.0.0
pytest==7.2.2
